requirements.txt
flask
gunicorn
lxml==4.4.1
beautifulsoup4==4.8.1
bs4==0.0.1
pandas==0.23.0
requests==2.22.0
line-bot-sdk
urllib3==1.25.7
matplotlib
numpy
pyimgur